still playing around with the navigation bar
